refactor(notifier): split edit and create into pages
